tl;dr: New engine goes brrrrrrr
A new engine for prediction movements has been implemented.  This engine tries to match multiple inputs to multiple movements to try and find the best combination.  In the future, the best predictions will have to be implemented first for optimization, such as weighting knockback in terms of the player's ping.  Right now, as the current implementation is going, cheaters could wait infinitely for knockback to apply (ignored kb), which theoretically would eventually crash the server.  Additionally, two similar knockbacks may be compared to one another, when the first one HAD to be applied before the second one, since the first one had to be applied.  Small knockback values may have to be dropped?  One issue with this is that previous checks could influence checks after itself.  Collision will now only run when it has to run, which is a big optimization.  The server no longer runs 18 collision checks per player per tick.  I don't see a good way to avoid at least one collision lookup.  If the predicted velocity is close enough to the actual lookup, I could reduce the number of collisions from two to one and just use the player's knockback.  I was planning on doing this anyways so that the previous check does not influence the next check, which helps a lot with dealing with bad predictions and other false positives, such as due to ping affecting blocks under the player not registering.  Initial transformation for supporting ladders and player knockback, which is implemented very similarly.  This still needs a lot of work in terms of optimization and definitely finishing everything other than just plain movements.  Movements in water, lava, and elytra still use the old engine.