Migrate to concurrent safe queue system.
Let's say the player moved from A -> B -> C.  The previous code added this to a singular queue for things to run, meaning that another thread could pick up a player's movement before the other thread finished the previous movement and check:
A -> C

This is a huge error and will cause false positives and a ton of other issues.  I solved this issue by adding the number of queue'd items to player data storage.  If there are queued items it adds them to a concurrent safe queue, which is first in first out.  If there aren't it adds the next item to the thread pool.  This queue is done with a ConcurrentHashMap so that their next item specific to each player can be added to the queue.

tl;dr: The queue for the thread pool can only hold or run one item for each player at a time.

f you see something wrong please tell me, as this is my own design to solve this issue and I haven't seen similar solutions elsewhere.  It took a few hours to design this.  I tried breaking my system by pausing with a debugger in a few places but I can't, and in theory everything seems multithreaded safe.

:)