Convert boats to a prediction engine - allow boats to take knockback
Thank you mojang for applying knockback into boats instead of allowing relative vehicle teleports or another solution that doesn't involve a hack of sending (0, 2.7, 0) and then (0, 0, 0) knockback when the player's boat is on top of a bubble column.  Oh well, at least we support plugins applying knockback to vehicles now.