Experimental pure uncertainty sneaking implementation
Sneaking was written incorrectly and varies up to 0.1 blocks per axis based on the hidden actual player speed.  It's better just to be uncertain here :(

I may attempt to rewrite this code later, but honestly I doubt it's worth trying to fix for a long time compared to other issues.  Yes, you can do illegitimate movement, but it's situation and shouldn't be an advantage, so what's even the point of trying to fix a movement mechanic that varies by 0.1 per axis...