Experimental pure uncertainty firework implementation
Fireworks are written quite bad in this terrible game.  They vary between ticks so the exact number of fireworks is unknown :(.  A single firework can tick twice in one tick, and no times the next.  With different look vectors, this really fucks up movement to the point of it being unpredictable.  I pointed this out on reddit when mojang was asking for feedback on fireworks and got downvoted.  Thanks Mojang (and /r/minecraft)